requirements.txt
boto3==1.34.162
botocore==1.34.162
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
